user and roles interface modification
git-svn-id: http://plugins.svn.wordpress.org/sliding-youtube-gallery/trunk@654602 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ js/core/lib/syg.lib.admin.js
@@ -200,7 +200,11 @@ jQuery.noConflict();
 						/* html = html + '<a href="#" onclick="javascript: jQuery.showDetails(\''+ val.id + '\');"><img src="../wp-content/plugins/sliding-youtube-gallery/img/ui/admin/details.png" title=""><img></a>'; */
 						html = html + '<a href="../wp-content/plugins/sliding-youtube-gallery/views/preview.php?id=' + val.id + '" class="iframe_' + val.id + '"><img src="../wp-content/plugins/sliding-youtube-gallery/img/ui/admin/preview.png" title="preview gallery"/></a>';
 						html = html + '<a href="?page=syg-manage-galleries&action=edit&id=' + val.id + '"><img src="../wp-content/plugins/sliding-youtube-gallery/img/ui/admin/edit.png" title="edit gallery"/></a>';
-						html = html + '<a href="#" onclick="javascript: jQuery.deleteGallery(\''+ val.id + '\');"><img src="../wp-content/plugins/sliding-youtube-gallery/img/ui/admin/delete.png" title="delete gallery"/></a>';
+						alert ($.cookie('syg-role'));
+						if ($.cookie('syg-role') == 'Administrator' || $.cookie('syg-role') == 'Editor') {
+							html = html + '<a href="#" onclick="javascript: jQuery.deleteGallery(\''+ val.id + '\');"><img src="../wp-content/plugins/sliding-youtube-gallery/img/ui/admin/delete.png" title="delete gallery"/></a>';
+						}
+						
 						if (val.cacheOn == 1) {
 							html = html + '<a href="#" onclick="javascript: jQuery.cacheGallery(\''+ val.id + '\');"><img src="../wp-content/plugins/sliding-youtube-gallery/img/ui/admin/cache.png" title="cache gallery"/></a>';
 						}
